backend/requirements.txt
Django>=4.2.11,<5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-decouple==3.8
Pillow==11.3.0
pytest==7.4.3
pytest-django==4.7.0
gunicorn==21.2.0
dj-database-url==2.1.0
# psycopg2-binary==2.9.9
django-filter==23.5